+++ ci4/app/Database/Migrations/2024-04-02-150935_AddWebhookHttpMethod.php
@@ -1,0 +1,24 @@
+<?php namespace App\Database\Migrations;
+
+use CodeIgniter\Database\Migration;
+use Config\Database;
+use OrmExtension\Migration\ColumnTypes;
+use OrmExtension\Migration\Table;
+
+class AddWebhookHttpMethod extends Migration {
+
+    public function up() {
+        Table::init('webhooks')
+            ->column('http_method', ColumnTypes::VARCHAR_63);
+
+        Database::connect()
+            ->table('webhooks')
+            ->set('http_method', 'post')
+            ->update();
+    }
+
+    public function down() {
+
+    }
+
+}